Add const correctness to ListCompareFxn
The arguments to ListCompareFxn really ought to be const, as compares ought
not modify what they're comparing.  This adds const to the function type,
and propagates the change down into the type-specific comparisons.

While I was in there, I also applied additional local const to each of the
compare implementations.