Let compiler compute sort registry length
Redefine SortRegistry so that the compiler can compute the length of the
sort registry, rather than relying on the human to update it.  This adds a
level of indirection; however, this data structure is not performance
critical.

Previously, SortRegistry had a flexible array member for the array of
registry entries.  It also has a length field that needs to match the
number of entries in that array.  C doesn't provide a straightforward way
to initialize the length field based on the actual length of the flexible
array member.  I could probably come up with something clever, but that's
not really called for here.

So, this change just replaces the flexible array member with an actual
pointer.  I now declare the array separately, and hook its pointer and
length into the SortRegistry structure.  This ensures that the length field
always reflects the length of the SortRegistry entry[] array.