Fold both benchmarks into `benchmark`
The last several commits generalized the benchmark driver code in
`benchmark_int64.c`, so that it ultimately did not directly depend on
any details of Int64ListNode.  Rather, it relied on an outside struct
to provide the mechanisms and details it needed.

This commit takes the final step, and upgrades `benchmark_int64.c` to
simply `benchmark.c`, with a single command-line argument to select
between Int64ListNode and CachelineListNode.  At the same time, it removes
the now-obsolete `benchmark_cacheline.c`.

    benchmark int64        Runs the Int64ListNode benchmark.
    benchmark cacheline    Runs the CachelineListNode benchmark.

A subsequent commit will update the documentation.